+++ src/languages/services/lexCompletions.ts
@@ -25,13 +25,15 @@ export function doLEXCompletion(document: TextDocument, position: Position, lexD
         return [];
     }
 
+    const word = document.getText(document.getWordRangeAtPosition(position)).toUpperCase();
+
     const line = document.lineAt(position.line).text.substring(0, position.character);
     const result: CompletionItem[] = [];
     if (offset < lexDocument.rulesRange[0]) {
         // if before rules zone, definition need to be on the right
         const ok = line.match(/^\w+.*({\w*}?)+/);
         if (ok) {
-            Object.keys(lexDocument.defines).forEach((key) => {
+            Object.keys(lexDocument.defines).filter(t=>t.toUpperCase().startsWith(word)).forEach((key) => {
                 const completion = new CompletionItem(key);
                 completion.detail = "definition";
                 completion.kind = CompletionItemKind.Class;
@@ -42,7 +44,7 @@ export function doLEXCompletion(document: TextDocument, position: Position, lexD
         const res = line.match(/^[^\s]*(?:{\w*}?)+$/);
         if (res) {
             if (res[0].length >= position.character) {
-                Object.keys(lexDocument.defines).forEach((key) => {
+                Object.keys(lexDocument.defines).filter(t=>t.toUpperCase().startsWith(word)).forEach((key) => {
                     const completion = new CompletionItem(key);
                     completion.detail = "definition";
                     completion.kind = CompletionItemKind.Class;
@@ -51,7 +53,7 @@ export function doLEXCompletion(document: TextDocument, position: Position, lexD
             }
         } else {
             if (line.match(/^<[\w,]*>[^\s]*(?:{\w*}?)+$/)) {
-                Object.keys(lexDocument.defines).forEach((key) => {
+                Object.keys(lexDocument.defines).filter(t=>t.toUpperCase().startsWith(word)).forEach((key) => {
                     const completion = new CompletionItem(key);
                     completion.detail = "definition";
                     completion.kind = CompletionItemKind.Class;
@@ -59,7 +61,7 @@ export function doLEXCompletion(document: TextDocument, position: Position, lexD
                 })
             } else if (line.match(/^<[\w,]*$/)) { // TODO: fix completion for {} after <>
 
-                Object.keys(lexDocument.states).forEach((key) => {
+                Object.keys(lexDocument.states).filter(t=>t.toUpperCase().startsWith(word)).forEach((key) => {
                     const completion = new CompletionItem(key);
                     completion.detail = "initial state";
                     completion.kind = CompletionItemKind.Class;