+++ src/languages/services/yaccCompletions.ts
@@ -27,6 +27,10 @@ export function doYACCComplete(document: TextDocument, position: Position, yaccD
         return [];
     }
 
+    const word = document.getText(document.getWordRangeAtPosition(position)).toUpperCase();
+    // this is used to match the completion items before we feed them out
+    // if we return an empty list, VSCode will fall back to the default 'abc' completions, which is what we want
+
     const node = yaccDocument.getNodeByOffset(offset);
     if (node === undefined) {
         return [];
@@ -39,7 +43,7 @@ export function doYACCComplete(document: TextDocument, position: Position, yaccD
         case NodeType.Type:
             if (node.typeOffset && offset > node.typeOffset) {
                 if (!node.typeEnd || offset <= node.typeEnd) {
-                    Object.keys(yaccDocument.types).forEach((type) => {
+                    Object.keys(yaccDocument.types).filter(t=>t.toUpperCase().startsWith(word)).forEach((type) => {
                         completion = new CompletionItem(type)
                         completion.detail = "type"
                         completion.kind = CompletionItemKind.TypeParameter;
@@ -49,7 +53,7 @@ export function doYACCComplete(document: TextDocument, position: Position, yaccD
                 }
             }
             if (node.nodeType === NodeType.Type)
-                Object.keys(yaccDocument.symbols).forEach((symbol) => {
+                Object.keys(yaccDocument.symbols).filter(t=>t.toUpperCase().startsWith(word)).forEach((symbol) => {
                     completion = new CompletionItem(symbol)
                     completion.detail = "user defined non-terminal";
                     completion.kind = CompletionItemKind.Class;
@@ -57,19 +61,19 @@ export function doYACCComplete(document: TextDocument, position: Position, yaccD
                 });
             break;
         case NodeType.Rule:
-            Object.keys(yaccDocument.symbols).forEach((symbol) => {
+            Object.keys(yaccDocument.symbols).filter(t=>t.toUpperCase().startsWith(word)).forEach((symbol) => {
                 completion = new CompletionItem(symbol)
                 completion.detail = "user defined non-terminal";
                 completion.kind = CompletionItemKind.Class;
                 result.push(completion);
             });
-            Object.keys(yaccDocument.tokens).forEach((token) => {
+            Object.keys(yaccDocument.tokens).filter(t=>t.toUpperCase().startsWith(word)).forEach((token) => {
                 completion = new CompletionItem(token)
                 completion.detail = "user defined token";
                 completion.kind = CompletionItemKind.Field;
                 result.push(completion);
             });
-            Object.keys(predefined).forEach(key => {
+            Object.keys(predefined).filter(t=>t.toUpperCase().startsWith(word)).forEach(key => {
                 completion = new CompletionItem(key)
                 completion.detail = "predefined symbol";
                 completion.kind = CompletionItemKind.Method;